Rename scaling to migration
