+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/java/org/apache/shardingsphere/migration/distsql/handler/update/ResetMigrationUpdater.java
@@ -15,27 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.migration.distsql.handler.handler.update;
+package org.apache.shardingsphere.migration.distsql.handler.update;
 
 import org.apache.shardingsphere.data.pipeline.api.RuleAlteredJobAPIFactory;
 import org.apache.shardingsphere.data.pipeline.api.RuleAlteredJobAPI;
 import org.apache.shardingsphere.infra.distsql.update.RALUpdater;
-import org.apache.shardingsphere.migration.distsql.statement.ApplyMigrationStatement;
+import org.apache.shardingsphere.migration.distsql.statement.ResetMigrationStatement;
 
 /**
- * Apply migration updater.
+ * Reset scaling updater.
  */
-public final class ApplyMigrationUpdater implements RALUpdater<ApplyMigrationStatement> {
+public final class ResetMigrationUpdater implements RALUpdater<ResetMigrationStatement> {
     
     private static final RuleAlteredJobAPI RULE_ALTERED_JOB_API = RuleAlteredJobAPIFactory.getInstance();
     
     @Override
-    public void executeUpdate(final ApplyMigrationStatement sqlStatement) {
-        RULE_ALTERED_JOB_API.switchClusterConfiguration(sqlStatement.getJobId());
+    public void executeUpdate(final ResetMigrationStatement sqlStatement) {
+        RULE_ALTERED_JOB_API.reset(sqlStatement.getJobId());
     }
     
     @Override
     public String getType() {
-        return ApplyMigrationStatement.class.getName();
+        return ResetMigrationStatement.class.getName();
     }
 }
