+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/java/org/apache/shardingsphere/migration/distsql/handler/update/RestoreMigrationSourceWritingUpdater.java
@@ -15,27 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.migration.distsql.handler.handler.update;
+package org.apache.shardingsphere.migration.distsql.handler.update;
 
 import org.apache.shardingsphere.data.pipeline.api.RuleAlteredJobAPIFactory;
 import org.apache.shardingsphere.data.pipeline.api.RuleAlteredJobAPI;
 import org.apache.shardingsphere.infra.distsql.update.RALUpdater;
-import org.apache.shardingsphere.migration.distsql.statement.StopMigrationStatement;
+import org.apache.shardingsphere.migration.distsql.statement.RestoreMigrationSourceWritingStatement;
 
 /**
- * Stop scaling updater.
+ * Restore scaling source writing updater.
  */
-public final class StopMigrationUpdater implements RALUpdater<StopMigrationStatement> {
+public final class RestoreMigrationSourceWritingUpdater implements RALUpdater<RestoreMigrationSourceWritingStatement> {
     
     private static final RuleAlteredJobAPI RULE_ALTERED_JOB_API = RuleAlteredJobAPIFactory.getInstance();
     
     @Override
-    public void executeUpdate(final StopMigrationStatement sqlStatement) {
-        RULE_ALTERED_JOB_API.stop(sqlStatement.getJobId());
+    public void executeUpdate(final RestoreMigrationSourceWritingStatement sqlStatement) {
+        RULE_ALTERED_JOB_API.restoreClusterWriteDB(sqlStatement.getJobId());
     }
     
     @Override
     public String getType() {
-        return StopMigrationStatement.class.getName();
+        return RestoreMigrationSourceWritingStatement.class.getName();
     }
 }
