+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.query.DistSQLResultSet
@@ -23,11 +23,11 @@ org.apache.shardingsphere.sharding.distsql.handler.query.ShardingAuditorsQueryRe
 org.apache.shardingsphere.sharding.distsql.handler.query.ShardingTableNodesQueryResultSet
 org.apache.shardingsphere.sharding.distsql.handler.query.ShardingKeyGeneratorsQueryResultSet
 org.apache.shardingsphere.sharding.distsql.handler.query.DefaultShardingStrategyQueryResultSet
-org.apache.shardingsphere.migration.distsql.handler.handler.query.ShardingScalingRulesQueryResultSet
-org.apache.shardingsphere.migration.distsql.handler.handler.query.CheckMigrationQueryResultSet
-org.apache.shardingsphere.migration.distsql.handler.handler.query.ShowMigrationListQueryResultSet
-org.apache.shardingsphere.migration.distsql.handler.handler.query.ShowMigrationJobStatusQueryResultSet
-org.apache.shardingsphere.migration.distsql.handler.handler.query.ShowMigrationCheckAlgorithmsQueryResultSet
+org.apache.shardingsphere.migration.distsql.handler.query.ShardingScalingRulesQueryResultSet
+org.apache.shardingsphere.migration.distsql.handler.query.CheckMigrationQueryResultSet
+org.apache.shardingsphere.migration.distsql.handler.query.ShowMigrationListQueryResultSet
+org.apache.shardingsphere.migration.distsql.handler.query.ShowMigrationJobStatusQueryResultSet
+org.apache.shardingsphere.migration.distsql.handler.query.ShowMigrationCheckAlgorithmsQueryResultSet
 org.apache.shardingsphere.sharding.distsql.handler.query.UnusedShardingAlgorithmsQueryResultSet
 org.apache.shardingsphere.sharding.distsql.handler.query.UnusedShardingKeyGeneratorsQueryResultSet
 org.apache.shardingsphere.sharding.distsql.handler.query.ShardingTableRulesUsedAlgorithmQueryResultSet
