+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.update.RALUpdater
@@ -15,10 +15,10 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.migration.distsql.handler.handler.update.StartMigrationUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.StopMigrationUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.ResetMigrationUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.DropMigrationUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.StopMigrationSourceWritingUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.RestoreMigrationSourceWritingUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.ApplyMigrationUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.StartMigrationUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.StopMigrationUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.ResetMigrationUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.DropMigrationUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.StopMigrationSourceWritingUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.RestoreMigrationSourceWritingUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.ApplyMigrationUpdater
