+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-handler/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.update.RuleDefinitionUpdater
@@ -35,7 +35,7 @@ org.apache.shardingsphere.sharding.distsql.handler.update.CreateShardingAuditorS
 org.apache.shardingsphere.sharding.distsql.handler.update.AlterShardingAuditorStatementUpdater
 org.apache.shardingsphere.sharding.distsql.handler.update.AlterDefaultShardingStrategyStatementUpdater
 org.apache.shardingsphere.sharding.distsql.handler.update.DropDefaultStrategyStatementUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.CreateShardingScalingRuleStatementUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.DropShardingScalingRuleStatementUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.EnableShardingScalingRuleStatementUpdater
-org.apache.shardingsphere.migration.distsql.handler.handler.update.DisableShardingScalingRuleStatementUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.CreateShardingScalingRuleStatementUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.DropShardingScalingRuleStatementUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.EnableShardingScalingRuleStatementUpdater
+org.apache.shardingsphere.migration.distsql.handler.update.DisableShardingScalingRuleStatementUpdater
