+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-statement/src/main/java/org/apache/shardingsphere/migration/distsql/statement/ShowMigrationStatusStatement.java
@@ -17,10 +17,16 @@
 
 package org.apache.shardingsphere.migration.distsql.statement;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.distsql.parser.statement.ral.scaling.QueryableScalingRALStatement;
 
 /**
- * Show scaling check algorithms statement.
+ * Show migration status statement.
  */
-public final class ShowScalingCheckAlgorithmsStatement extends QueryableScalingRALStatement {
+@RequiredArgsConstructor
+@Getter
+public final class ShowMigrationStatusStatement extends QueryableScalingRALStatement {
+    
+    private final String jobId;
 }
